fix: Circular dependency in Edit Members
